+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Source, Metadata and Entity structs are now truncated for `inspect`, omitting data and parsed XML, for easier debugging
 - Source, Metadata and Entity structs can now be interpolated and printed in strings, showing a type and unique URI 
   (format may change in future releases)
+- `Smee.Fetch.warm/2` will download a list of sources concurrently and warm the HTTP cache
 
 ## Improvements
 - A few more metadata namespaces have been added